@@ -1,10 +1,13 @@
 {
   "name": "@webpack-cli/configtest",
-  "version": "1.2.0",
+  "version": "2.1.1",
   "description": "Validate a webpack configuration.",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "license": "MIT",
+  "engines": {
+    "node": ">=14.15.0"
+  },
   "publishConfig": {
     "access": "public"
   },
@@ -17,8 +20,8 @@
     "lib"
   ],
   "peerDependencies": {
-    "webpack": "4.x.x || 5.x.x",
-    "webpack-cli": "4.x.x"
+    "webpack": "5.x.x",
+    "webpack-cli": "5.x.x"
   },
-  "gitHead": "20882d463450d010bb76e0824fe555e9785e9561"
+  "gitHead": "e879ce4ef91a9a89ca5ef74f533391cef5ba009d"
 }
