@@ -1,6 +1,6 @@
 {
   "name": "schema-utils",
-  "version": "2.7.1",
+  "version": "4.2.0",
   "description": "webpack Validation Utils",
   "license": "MIT",
   "repository": "webpack/schema-utils",
@@ -14,7 +14,7 @@
   "main": "dist/index.js",
   "types": "declarations/index.d.ts",
   "engines": {
-    "node": ">= 8.9.0"
+    "node": ">= 12.13.0"
   },
   "scripts": {
     "start": "npm run build -- -w",
@@ -24,51 +24,54 @@
     "build:code": "cross-env NODE_ENV=production babel src -d dist --copy-files",
     "build": "npm-run-all -p \"build:**\"",
     "commitlint": "commitlint --from=master",
-    "security": "npm audit",
-    "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css,ts}\" --list-different",
+    "security": "npm audit --production",
+    "fmt:check": "prettier \"{**/*,*}.{js,json,md,yml,css,ts}\" --list-different",
     "lint:js": "eslint --cache .",
     "lint:types": "tsc --pretty --noEmit",
-    "lint": "npm-run-all -l -p \"lint:**\"",
+    "lint": "npm-run-all lint:js lint:types fmt:check",
+    "fmt": "npm run fmt:check -- --write",
+    "fix:js": "npm run lint:js -- --fix",
+    "fix": "npm-run-all fix:js fmt",
     "test:only": "cross-env NODE_ENV=test jest",
     "test:watch": "npm run test:only -- --watch",
     "test:coverage": "npm run test:only -- --collectCoverageFrom=\"src/**/*.js\" --coverage",
     "pretest": "npm run lint",
     "test": "npm run test:coverage",
-    "prepare": "npm run build",
-    "release": "standard-version",
-    "defaults": "webpack-defaults"
+    "prepare": "npm run build && husky install",
+    "release": "standard-version"
   },
   "files": [
     "dist",
     "declarations"
   ],
   "dependencies": {
-    "ajv": "^6.12.4",
-    "ajv-keywords": "^3.5.2",
-    "@types/json-schema": "^7.0.5"
+    "@types/json-schema": "^7.0.9",
+    "ajv": "^8.9.0",
+    "ajv-formats": "^2.1.1",
+    "ajv-keywords": "^5.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.11.4",
-    "@babel/preset-env": "^7.11.0",
-    "@commitlint/cli": "^10.0.0",
-    "@commitlint/config-conventional": "^10.0.0",
-    "@webpack-contrib/defaults": "^6.3.0",
+    "@babel/cli": "^7.17.0",
+    "@babel/core": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@commitlint/cli": "^17.6.1",
+    "@commitlint/config-conventional": "^16.0.0",
     "@webpack-contrib/eslint-config-webpack": "^3.0.0",
-    "babel-jest": "^25.5.1",
-    "cross-env": "^6.0.3",
-    "del": "^5.1.0",
-    "del-cli": "^3.0.1",
-    "eslint": "^7.7.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.0",
-    "husky": "^4.2.5",
-    "jest": "^25.5.4",
-    "lint-staged": "^10.2.13",
+    "babel-jest": "^27.4.6",
+    "cross-env": "^7.0.3",
+    "del": "^6.0.0",
+    "del-cli": "^4.0.1",
+    "eslint": "^8.8.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.4",
+    "husky": "^7.0.4",
+    "jest": "^27.4.7",
+    "lint-staged": "^13.2.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.1",
-    "standard-version": "^9.0.0",
-    "typescript": "^4.0.2"
+    "prettier": "^2.5.1",
+    "standard-version": "^9.3.2",
+    "typescript": "^4.9.5",
+    "webpack": "^5.68.0"
   },
   "keywords": [
     "webpack"
