@@ -1,23 +1,21 @@
 {
   "name": "babel-loader",
-  "version": "8.3.0",
+  "version": "9.1.3",
   "description": "babel module loader for webpack",
   "files": [
     "lib"
   ],
   "main": "lib/index.js",
   "engines": {
-    "node": ">= 8.9"
+    "node": ">= 14.15.0"
   },
   "dependencies": {
-    "find-cache-dir": "^3.3.1",
-    "loader-utils": "^2.0.0",
-    "make-dir": "^3.1.0",
-    "schema-utils": "^2.6.5"
+    "find-cache-dir": "^4.0.0",
+    "schema-utils": "^4.0.0"
   },
   "peerDependencies": {
-    "@babel/core": "^7.0.0",
-    "webpack": ">=2"
+    "@babel/core": "^7.12.0",
+    "webpack": ">=5"
   },
   "devDependencies": {
     "@ava/babel": "^1.0.1",
@@ -25,9 +23,9 @@
     "@babel/core": "^7.19.6",
     "@babel/preset-env": "^7.19.4",
     "ava": "^3.13.0",
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-istanbul": "^6.0.0",
-    "babel-plugin-react-intl": "^8.2.15",
+    "babel-eslint": "^10.1.0",
+    "babel-plugin-istanbul": "^6.1.1",
+    "babel-plugin-react-intl": "^8.2.25",
     "cross-env": "^7.0.2",
     "eslint": "^7.13.0",
     "eslint-config-babel": "^9.0.0",
@@ -43,8 +41,8 @@
     "react-intl": "^5.9.4",
     "react-intl-webpack-plugin": "^0.3.0",
     "rimraf": "^3.0.0",
-    "semver": "7.3.2",
-    "webpack": "^5.34.0"
+    "semver": "7.5.2",
+    "webpack": "^5.74.0"
   },
   "scripts": {
     "clean": "rimraf lib/",
