@@ -1,6 +1,6 @@
 {
   "name": "commander",
-  "version": "7.2.0",
+  "version": "10.0.1",
   "description": "the complete solution for node.js command-line programs",
   "keywords": [
     "commander",
@@ -19,36 +19,48 @@
     "url": "https://github.com/tj/commander.js.git"
   },
   "scripts": {
-    "lint": "eslint index.js esm.mjs \"tests/**/*.js\"",
-    "typescript-lint": "eslint typings/*.ts tests/*.ts",
+    "lint": "npm run lint:javascript && npm run lint:typescript",
+    "lint:javascript": "eslint index.js esm.mjs \"lib/*.js\" \"tests/**/*.js\"",
+    "lint:typescript": "eslint typings/*.ts tests/*.ts",
     "test": "jest && npm run test-typings",
-    "test-esm": "node --experimental-modules ./tests/esm-imports-test.mjs",
+    "test-esm": "node ./tests/esm-imports-test.mjs",
     "test-typings": "tsd",
-    "typescript-checkJS": "tsc --allowJS --checkJS index.js --noEmit",
-    "test-all": "npm run test && npm run lint && npm run typescript-lint && npm run typescript-checkJS && npm run test-esm"
+    "typescript-checkJS": "tsc --allowJS --checkJS index.js lib/*.js --noEmit",
+    "test-all": "npm run test && npm run lint && npm run typescript-checkJS && npm run test-esm"
   },
-  "main": "./index.js",
   "files": [
     "index.js",
+    "lib/*.js",
     "esm.mjs",
     "typings/index.d.ts",
     "package-support.json"
   ],
   "type": "commonjs",
-  "dependencies": {},
+  "main": "./index.js",
+  "exports": {
+    ".": {
+      "types": "./typings/index.d.ts",
+      "require": "./index.js",
+      "import": "./esm.mjs"
+    },
+    "./esm.mjs": "./esm.mjs"
+  },
   "devDependencies": {
-    "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.20",
-    "@typescript-eslint/eslint-plugin": "^4.12.0",
-    "@typescript-eslint/parser": "^4.12.0",
-    "eslint": "^7.17.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-jest": "^24.1.3",
-    "jest": "^26.6.3",
-    "standard": "^16.0.3",
-    "ts-jest": "^26.5.1",
-    "tsd": "^0.14.0",
-    "typescript": "^4.1.2"
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
+    "eslint": "^8.30.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-config-standard-with-typescript": "^24.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.1.7",
+    "eslint-plugin-n": "^15.6.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3",
+    "tsd": "^0.25.0",
+    "typescript": "^4.9.4"
   },
   "types": "typings/index.d.ts",
   "jest": {
@@ -62,7 +74,7 @@
     ]
   },
   "engines": {
-    "node": ">= 10"
+    "node": ">=14"
   },
   "support": true
 }
