@@ -1,6 +1,6 @@
 {
   "name": "webpack-cli",
-  "version": "4.10.0",
+  "version": "5.1.4",
   "description": "CLI for webpack & friends",
   "license": "MIT",
   "repository": {
@@ -18,7 +18,7 @@
   },
   "main": "./lib/index.js",
   "engines": {
-    "node": ">=10.13.0"
+    "node": ">=14.15.0"
   },
   "keywords": [
     "webpack",
@@ -35,26 +35,27 @@
   ],
   "dependencies": {
     "@discoveryjs/json-ext": "^0.5.0",
-    "@webpack-cli/configtest": "^1.2.0",
-    "@webpack-cli/info": "^1.5.0",
-    "@webpack-cli/serve": "^1.7.0",
+    "@webpack-cli/configtest": "^2.1.1",
+    "@webpack-cli/info": "^2.0.2",
+    "@webpack-cli/serve": "^2.0.5",
     "colorette": "^2.0.14",
-    "commander": "^7.0.0",
+    "commander": "^10.0.1",
     "cross-spawn": "^7.0.3",
+    "envinfo": "^7.7.3",
     "fastest-levenshtein": "^1.0.12",
     "import-local": "^3.0.2",
-    "interpret": "^2.2.0",
-    "rechoir": "^0.7.0",
+    "interpret": "^3.1.1",
+    "rechoir": "^0.8.0",
     "webpack-merge": "^5.7.3"
   },
+  "devDependencies": {
+    "@types/envinfo": "^7.8.1"
+  },
   "peerDependencies": {
-    "webpack": "4.x.x || 5.x.x"
+    "webpack": "5.x.x"
   },
   "peerDependenciesMeta": {
     "@webpack-cli/generators": {
-      "optional": true
-    },
-    "@webpack-cli/migrate": {
       "optional": true
     },
     "webpack-bundle-analyzer": {
@@ -64,5 +65,5 @@
       "optional": true
     }
   },
-  "gitHead": "20882d463450d010bb76e0824fe555e9785e9561"
+  "gitHead": "e07f0e58df103011435524d757102534b75a6796"
 }
