@@ -10,21 +9,172 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.12.5",
-    "babel-loader": "^8.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "webpack": "^5.4.0",
-    "webpack-cli": "^4.2.0"
-  },
   "dependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.61",
-    "react-router-dom": "^5.2.0"
-  }
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.16",
+    "@mui/material": "^5.14.17",
+    "acorn": "^8.10.0",
+    "acorn-import-assertions": "^1.9.0",
+    "ajv": "^6.12.6",
+    "ajv-keywords": "^3.5.2",
+    "ansi-styles": "^3.2.1",
+    "babel-plugin-polyfill-corejs2": "^0.4.6",
+    "babel-plugin-polyfill-corejs3": "^0.8.6",
+    "babel-plugin-polyfill-regenerator": "^0.5.3",
+    "big.js": "^5.2.2",
+    "browserslist": "^4.22.1",
+    "buffer-from": "^1.1.2",
+    "caniuse-lite": "^1.0.30001546",
+    "chalk": "^2.4.2",
+    "chrome-trace-event": "^1.0.3",
+    "clone-deep": "^4.0.1",
+    "clsx": "^1.2.1",
+    "color-convert": "^1.9.3",
+    "color-name": "^1.1.3",
+    "colorette": "^2.0.20",
+    "commander": "^2.20.3",
+    "commondir": "^1.0.1",
+    "convert-source-map": "^2.0.0",
+    "core-js-compat": "^3.33.2",
+    "cross-spawn": "^7.0.3",
+    "css-vendor": "^2.0.8",
+    "csstype": "^3.1.2",
+    "debug": "^4.3.4",
+    "dom-helpers": "^5.2.1",
+    "electron-to-chromium": "^1.4.544",
+    "emojis-list": "^3.0.0",
+    "enhanced-resolve": "^5.15.0",
+    "envinfo": "^7.10.0",
+    "es-module-lexer": "^1.3.1",
+    "escalade": "^3.1.1",
+    "escape-string-regexp": "^1.0.5",
+    "eslint-scope": "^5.1.1",
+    "esrecurse": "^4.3.0",
+    "estraverse": "^4.3.0",
+    "esutils": "^2.0.3",
+    "events": "^3.3.0",
+    "fast-deep-equal": "^3.1.3",
+    "fast-json-stable-stringify": "^2.1.0",
+    "fastest-levenshtein": "^1.0.16",
+    "find-cache-dir": "^3.3.2",
+    "find-up": "^4.1.0",
+    "gensync": "^1.0.0-beta.2",
+    "glob-to-regexp": "^0.4.1",
+    "globals": "^11.12.0",
+    "graceful-fs": "^4.2.11",
+    "has": "^1.0.4",
+    "has-flag": "^4.0.0",
+    "history": "^4.10.1",
+    "hoist-non-react-statics": "^3.3.2",
+    "hyphenate-style-name": "^1.0.4",
+    "import-local": "^3.1.0",
+    "interpret": "^2.2.0",
+    "is-core-module": "^2.13.0",
+    "is-in-browser": "^1.1.3",
+    "is-plain-object": "^2.0.4",
+    "isarray": "^0.0.1",
+    "isexe": "^2.0.0",
+    "isobject": "^3.0.1",
+    "jest-worker": "^27.5.1",
+    "js-tokens": "^4.0.0",
+    "jsesc": "^2.5.2",
+    "json-parse-even-better-errors": "^2.3.1",
+    "json-schema-traverse": "^0.4.1",
+    "json5": "^2.2.3",
+    "jss": "^10.10.0",
+    "jss-plugin-camel-case": "^10.10.0",
+    "jss-plugin-default-unit": "^10.10.0",
+    "jss-plugin-global": "^10.10.0",
+    "jss-plugin-nested": "^10.10.0",
+    "jss-plugin-props-sort": "^10.10.0",
+    "jss-plugin-rule-value-function": "^10.10.0",
+    "jss-plugin-vendor-prefixer": "^10.10.0",
+    "kind-of": "^6.0.3",
+    "loader-runner": "^4.3.0",
+    "loader-utils": "^2.0.4",
+    "locate-path": "^5.0.0",
+    "lodash.debounce": "^4.0.8",
+    "loose-envify": "^1.4.0",
+    "lru-cache": "^5.1.1",
+    "make-dir": "^3.1.0",
+    "merge-stream": "^2.0.0",
+    "mime-db": "^1.52.0",
+    "mime-types": "^2.1.35",
+    "ms": "^2.1.2",
+    "neo-async": "^2.6.2",
+    "node-releases": "^2.0.13",
+    "object-assign": "^4.1.1",
+    "p-limit": "^2.3.0",
+    "p-locate": "^4.1.0",
+    "p-try": "^2.2.0",
+    "path-exists": "^4.0.0",
+    "path-key": "^3.1.1",
+    "path-parse": "^1.0.7",
+    "path-to-regexp": "^1.8.0",
+    "picocolors": "^1.0.0",
+    "pkg-dir": "^4.2.0",
+    "popper.js": "^1.16.1-lts",
+    "prop-types": "^15.8.1",
+    "punycode": "^2.3.0",
+    "randombytes": "^2.1.0",
+    "react-is": "^17.0.2",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^6.18.0",
+    "react-transition-group": "^4.4.5",
+    "rechoir": "^0.7.1",
+    "regenerate": "^1.4.2",
+    "regenerate-unicode-properties": "^10.1.1",
+    "regenerator-runtime": "^0.14.0",
+    "regenerator-transform": "^0.15.2",
+    "regexpu-core": "^5.3.2",
+    "regjsparser": "^0.9.1",
+    "resolve": "^1.22.6",
+    "resolve-cwd": "^3.0.0",
+    "resolve-from": "^5.0.0",
+    "resolve-pathname": "^3.0.0",
+    "safe-buffer": "^5.2.1",
+    "scheduler": "^0.20.2",
+    "schema-utils": "^3.3.0",
+    "semver": "^6.3.1",
+    "serialize-javascript": "^6.0.1",
+    "shallow-clone": "^3.0.1",
+    "shebang-command": "^2.0.0",
+    "shebang-regex": "^3.0.0",
+    "source-map": "^0.6.1",
+    "source-map-support": "^0.5.21",
+    "supports-color": "^8.1.1",
+    "supports-preserve-symlinks-flag": "^1.0.0",
+    "tapable": "^2.2.1",
+    "terser": "^5.21.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "tiny-invariant": "^1.3.1",
+    "tiny-warning": "^1.0.3",
+    "to-fast-properties": "^2.0.0",
+    "unicode-canonical-property-names-ecmascript": "^2.0.0",
+    "unicode-match-property-ecmascript": "^2.0.0",
+    "unicode-match-property-value-ecmascript": "^2.1.0",
+    "unicode-property-aliases-ecmascript": "^2.1.0",
+    "update-browserslist-db": "^1.0.13",
+    "uri-js": "^4.4.1",
+    "value-equal": "^1.0.1",
+    "watchpack": "^2.4.0",
+    "webpack-merge": "^5.9.0",
+    "webpack-sources": "^3.2.3",
+    "which": "^2.0.2",
+    "wildcard": "^2.0.1",
+    "yallist": "^3.1.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.23.3",
+    "@babel/preset-env": "^7.23.3",
+    "@babel/preset-react": "^7.23.3",
+    "babel-loader": "^9.1.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  },
+  "description": ""
 }
