@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { TextField, Button, Grid, Typography } from "@material-ui/core";
+import { TextField, Button, Grid, Typography } from "@mui/material";
 import { Link } from "react-router-dom";
 
 export default class RoomJoinPage extends Component {
